bug-fixing(export-case): exlpode function export perfunction 25 lines refactoring same filter rename dictionary
Fixes #616
